+++ src/app/home/page.tsx
@@ -1,7 +1,7 @@
 import JohnjudImage from "@/assets/johnjud-logo-with-text-side.png";
 import Button from "@/components/Button";
 import Container from "@/components/Container";
-import Divider from "@/components/Divider/Divider";
+import Divider from "@/components/Divider";
 import Banner from "@/components/Main/Banner";
 import BigPetCardContainer from "@/components/Main/CardList/BigPetCardContainer";
 import MiniPetCardContainer from "@/components/Main/CardList/MiniPetCardContainer";
